Update README.md
Detecting age is a regression problem. I have used reLU as the activation function and mean absolute error as the loss function. Detecting gender is a classification problem. I have used Sigmoid as the activation function and Binary cross-entropy as the loss function. Tensorflow and OpenCV libraries are used for implementation.